backend/requirements.txt
# =========================

# BACKEND WEB

# =========================

flask==3.0.3
flask-restx==1.3.0
flask-cors==4.0.0
flask-bcrypt==1.0.1

# =========================

# BANCO DE DADOS

# =========================

sqlalchemy==2.0.29
pymysql==1.1.0

# =========================

# IA / RAG

# =========================

langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-groq==0.1.4
sentence-transformers==2.7.0
faiss-cpu==1.7.4

# =========================

# MODELOS HUGGINGFACE

# =========================

transformers==4.41.2
huggingface-hub==0.23.0
torch==2.2.2

# =========================

# PROCESSAMENTO DE TEXTO

# =========================

langchain-text-splitters==0.0.2

# =========================

# PDF

# =========================

PyPDF2==3.0.1
pypdf==4.2.0

# =========================

# UTILIDADES

# =========================

python-dotenv==1.0.1
pydantic==1.10.13
tqdm==4.67.3